+++ test/fixtures/tar_plan/_targets.R
@@ -1,0 +1,5 @@
+tarchetypes::tar_plan(
+	alpha = 1,
+	beta = alpha + 1,
+	tar_target(gamma, beta + 1)
+)